+++ CHANGELOG.md
@@ -14,6 +14,7 @@ focus on the user impact** rather than the actual changes made.
 ## Added
 * Achievement rule numbering. In the future this will be used to allow the user to enable/disable
   the rules according to their preference.
+* Added the `H2-shortest-subject-line` achievement.
 
 ## Changed
 ## Deprecated